add postgres custom docker image; clean up compose
the default `binami/postgres:14` image was updated to remove basic
commands like curl, and thus causing the bootstrap script to fail (see

This commit adds a very thin customization to the bitnamic's postgres
image by adding `RUN install_packages curl` to add curl to the image.
This resovles the issue.

The commit also cleans up the docker-compose file a bit with the
following changes:
- give the compose a name `bgpkit-broker`
- clean up service names, remove `bgpkit-broker-` prefix
- add custom postgres container and build